move run to postprocessing_runner.py
